requirements.txt
couchbase
python-dotenv
